32bit-fix

+++ src/Core/Attributes/Migration.php
@@ -7,7 +7,7 @@ use Attribute;
 #[Attribute(Attribute::TARGET_CLASS)]
 class Migration {
 	public function __construct(
-		public int $order,
+		public float $order,
 		public bool $shared=false,
 	) {
 	}
